pyproject.toml
[tool.poetry]
name = "random-image-api"
version = "0.1.0"
description = ""
authors = ["breno-jesus-fernandes <breno.de.jesus.fernandes@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "random_image_api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
pillow = "^10.0.1"
uvicorn = "^0.23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
